Add files via upload
In my master thesis, I tried a wheel legged robot. The primary objective of this my research is to initiate with the design of a widely employed linear controller, namely the Linear Quadratic Regulator (LQR), based on a linearized model of the robot. This baseline controller is then progressively enhanced through the implementation of either a model-based nonlinear controller or a robust adaptive control scheme. For computational simplification, the robot is modeled as a two-wheeled inverted pendulum, where the height of the robot’s center of mass is dynamically adjustable through changes in the angles of the knee & hip joints. Subsequently, an LQR controller is designed to facilitate the robot’s ability to maintain balanced while simultaneously achieving locomotion in various directions. The dynamic stability of the robot is rigorously analyzed, accounting for modeling uncertainties, sensor noise, and external disturbances. Following this, a model-based nonlinear computed torque controller is formulated for the wheel-legged robot. Parametric uncertainties inherent in the model, such as friction coefficients, inertia matrices, and center of mass position, generally degrade the performance of model-based motion controllers. To mitigate these performance losses, a hybrid control structure is proposed, consisting of an LQR feedback loop augmented with a feedforward computed torque control component. To further enhance the robustness of the motion controller against external disturbances and parametric uncertainties, the LQR controller is reinforced with an advanced ℒ1 adaptive controller. I modeled this robot in Gazebo environment and controled it with C++ programming language.